Added datamanager for restaurants. Moved dishes to data folder
