+++ app/src/main/java/com/example/feedme/CollectOrderRecyclerAdapter.kt
@@ -4,13 +4,10 @@ import android.content.Context
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageButton
 import android.widget.TextView
-import androidx.appcompat.view.menu.MenuView.ItemView
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-
-import com.example.feedme.data.Restaurant
+import com.example.feedme.data.Dishes
 
 //Todo byta dishes till restaurants när all data är ner
