+++ app/src/main/java/com/example/feedme/CustomerAdressDeliveryReceiyclerAdapter.kt
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.example.feedme.data.Dishes
 
 class CustomerAdressDeliveryReceiyclerAdapter (val context: Context,
 val dishes: List<Dishes>) :
