+++ app/src/main/java/com/example/feedme/DataManagerRestaurants.kt
@@ -1,0 +1,7 @@
+package com.example.feedme
+
+import com.example.feedme.data.Restaurant
+
+object DataManagerRestaurants {
+    val restaurants = mutableListOf<Restaurant>()
+}
