+++ app/src/main/java/com/example/feedme/FoodViewRecyclerAdapter.kt
@@ -2,7 +2,6 @@ package com.example.feedme
 
 import android.content.Context
 import android.content.Intent
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -12,8 +11,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.core.view.isInvisible
 import androidx.recyclerview.widget.RecyclerView
-import androidx.recyclerview.widget.RecyclerView.ViewHolder
-import kotlin.math.log
+import com.example.feedme.data.Dishes
 
 lateinit var dishes: Dishes
