+++ app/src/main/java/com/example/feedme/InfoRestaurantActivity.kt
@@ -94,4 +94,5 @@ class InfoRestaurantActivity : AppCompatActivity() {
 
         return toReturn.trim()
     }
+
 }
